requirements.txt
# Python 3.12+ Compatible Dependencies

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Vector Databases
pymilvus>=2.4.0
faiss-cpu>=1.8.0

# Deep Learning
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0
imagehash>=4.3.0

# Data Processing
numpy>=1.24.0,<3.0  # Updated for Python 3.12 compatibility
pandas>=2.0.0
scikit-learn>=1.3.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# API & Validation
pydantic>=2.5.0
pydantic-settings>=2.0.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.0.0
tqdm>=4.65.0
colorama>=0.4.6
redis>=5.0.0
types-redis>=4.6.0

# Logging
loguru>=0.7.0


